Create a tensor histogram summary
Summary:
This commit introduces a new-style summary for histograms, which new
code should prefer over `tf.summary.histogram`. The new summary is not
privileged and is built on top of `tensor_summary`.

I've updated the histogram dashboard to support both the old and new
forms.

Once `display_name` and `description` are implemented, we can start
using those in the API, backend, and frontend. Doing so will give us the
opportunity to strip out the ugly `.../histogram_summary` from the tag
on the frontend. (I could have done this in this commit, but wanted to
keep it reasonably sized.)

Note: we don't have a way to run the `histogramTests` on the frontend,
and per #174 they don't pass anyway. I tried to make them valid but did
not run them and provide no guarantees!

Test Plan:
Generate test data by changing the demo's log directory to
`/tmp/new_histograms_demo` and re-running the demo. (Make sure to keep
your old data for comparison.) Change the event accumulator's size
guidance for tensors to `50` to match that of the old-style histograms.
Then, launch TensorBoard and make sure that the histograms render
correctly, both new and old. (There may be some minor bucketing
differences because the old-style histograms create their buckets in a
different way and also have different `min` and `max` than new-style
histograms, but the representations should be qualitatively the same.)

Screenshot:
![Example screenshot (new and old data).](https://user-images.githubusercontent.com/4317806/28437087-3763d16c-6d4e-11e7-9740-73ad563a4af1.png)

Also, run tests!

wchargin-branch: tensor-histogram-summary